+++ NEWS.md
@@ -1,5 +1,12 @@
 # constructive (development version)
 
+* Set `options(constructive_print = <character>)` where `<character>` is a vector
+    of strings among `"console"`, `"clipboard"` and `"reprex"`. The default
+    behavior is "console". If you use `"clipboard"`, with or without `"reprex"`,
+    the clipboard will contain only the code. If you use `"reprex"` without
+    `"clipboard"` the clipboard will still be used to receive the reprex,
+    can be tweaked with `options(reprex.clipboard = FALSE)`
+
 # constructive 0.2.0
 
 * We don't use {styler} anymore, performance is sensibly enhanced as